Correção, livos/usuarios são listados apenas para administradores

+++ src/main/java/br/edu/ifg/projetoweb/controller/LivroController.java
@@ -12,64 +12,78 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import br.edu.ifg.projetoweb.DAO.LivroDAO;
+import br.edu.ifg.projetoweb.DAO.UsuarioDAO;
 import br.edu.ifg.projetoweb.core.ProjectHttpServlet;
 import br.edu.ifg.projetoweb.model.Livro;
+import br.edu.ifg.projetoweb.utils.Sessao;
 
-
-
-@WebServlet (name = "ListarLivro", value = "/livros/*")
-public class LivroController  extends ProjectHttpServlet {
+@WebServlet(name = "ListarLivro", value = "/livros/*")
+public class LivroController extends ProjectHttpServlet {
 
 	private static final long serialVersionUID = 1L;
 	private List<Livro> livros = new ArrayList<>();
 
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws IOException, ServletException {
 		Connection connection = (Connection) request.getAttribute("connection");
 		LivroDAO livroDAO = new LivroDAO(connection);
+		int idUsuario = Sessao.getUsuarioId(request);
+
+		UsuarioDAO usuarioDAO = new UsuarioDAO(connection);
 
 		String pathInfo = request.getPathInfo();
-		
-		if(pathInfo == null || pathInfo.equals("/")) {
-			try {
-				livros = livroDAO.listarLivros();
-			}catch(SQLException e) {
-				e.printStackTrace();
+
+		if (pathInfo == null || pathInfo.equals("/")) {
+			if (usuarioDAO.isAdmin(idUsuario)) {
+				try {
+					livros = livroDAO.listarLivros();
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+				request.setAttribute("livros", livros);
 			}
-		
-		
-		request.setAttribute("livros", livros);
-		Router.listarLivros(request, response);
-		return;
-		}
-		
-		String[] splits = pathInfo.split("/");
-		
-		if(splits.length != 2) {
+			Router.listarLivros(request, response);
 			return;
 		}
-		if(splits[1].matches("cadastrar")) {
-			Router.cadastrarLivro(request, response);
-			return ;
-			
-		}
-		
-		if (splits[1].matches("home")) {
 
+		String[] splits = pathInfo.split("/");
+
+		if (splits.length != 2) {
+			return;
+		}
+
+		if (splits[1].matches("cadastrar")) {
+			if (usuarioDAO.isAdmin(idUsuario)) {
+				Router.cadastrarLivro(request, response);
+			} else {
+				response.sendRedirect("/projetoweb/livros");
+			}
+			return;
+		}
+
+		if (splits[1].matches("home")) {
 			Router.home(request, response);
 			return;
+		}
 
+		int id = Integer.parseInt(splits[1]);
+
+		if (usuarioDAO.isAdmin(idUsuario)) {
+			try {
+				Livro livro = livroDAO.buscarLivro(id);
+				request.setAttribute("livro", livro);
+				Router.atualizarLivro(request, response);
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+		} else {
+			response.sendRedirect("/projetoweb/livros");
 		}
-		int id = Integer.parseInt(splits[1]);	
-		try {
-			Livro livro = livroDAO.buscarLivro(id);
-			request.setAttribute("livro", livro);
-			Router.atualizarLivro(request, response);
-		}catch (Exception e) {
-			e.printStackTrace();
-		}
+
 	}
 
-	protected void doPut(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
+	protected void doPut(HttpServletRequest request, HttpServletResponse response)
+			throws IOException, ServletException {
 
 		Connection connection = (Connection) request.getAttribute("connection");
 		LivroDAO livroDAO = new LivroDAO(connection);
@@ -95,52 +109,61 @@ public class LivroController  extends ProjectHttpServlet {
 		response.sendRedirect("/projetoweb/livros");
 
 	}
+
 	protected void doDelete(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		
-			Connection connection = (Connection) request.getAttribute("connection");
-			LivroDAO livroDAO = new LivroDAO(connection);
-			List<Livro> livros = null;
-			int id = Integer.parseInt(request.getParameter("id"));
-			
+
+		Connection connection = (Connection) request.getAttribute("connection");
+		LivroDAO livroDAO = new LivroDAO(connection);
+		List<Livro> livros = null;
+
+		int id = Integer.parseInt(request.getParameter("id"));
+		int idUsuario = Sessao.getUsuarioId(request);
+
+		UsuarioDAO usuarioDAO = new UsuarioDAO(connection);
+
+		if (usuarioDAO.isAdmin(idUsuario)) {
 			livroDAO.removerLivro(id);
-			
-			try {
-				livros = livroDAO.listarLivros();
-			}catch(SQLException e) {
-				e.printStackTrace();
-			}
-			
-			
-			request.setAttribute("livros", livros);
-			Router.listarLivros(request, response);
-			
 		}
-	
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+		try {
+			livros = livroDAO.listarLivros();
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+
+		request.setAttribute("livros", livros);
+		Router.listarLivros(request, response);
+
+	}
+
+	protected void doPost(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
 		final Connection connection = (Connection) request.getAttribute("connection");
 		final LivroDAO livroDAO = new LivroDAO(connection);
 		final Livro livro = new Livro();
-		
-		livro.setNome(request.getParameter("nome"));
-		livro.setAutor(request.getParameter("autor"));
-		livro.setDescricao(request.getParameter("descricao"));
-		livro.setValor(Double.parseDouble(request.getParameter("valor")));
-		livro.setISBN(request.getParameter("isbn"));
-		livro.setQuantidade(Integer.parseInt(request.getParameter("quantidade")));
-		
-		
-		try {
-			livroDAO.inserirLivro(livro);
-			request.removeAttribute("connection");
-		}catch(SQLException e1) {
-			e1.printStackTrace();
+
+		int idUsuario = Sessao.getUsuarioId(request);
+		UsuarioDAO usuarioDAO = new UsuarioDAO(connection);
+
+		if (usuarioDAO.isAdmin(idUsuario)) {
+
+			livro.setNome(request.getParameter("nome"));
+			livro.setAutor(request.getParameter("autor"));
+			livro.setDescricao(request.getParameter("descricao"));
+			livro.setValor(Double.parseDouble(request.getParameter("valor")));
+			livro.setISBN(request.getParameter("isbn"));
+			livro.setQuantidade(Integer.parseInt(request.getParameter("quantidade")));
+
+			try {
+				livroDAO.inserirLivro(livro);
+				request.removeAttribute("connection");
+			} catch (SQLException e1) {
+				e1.printStackTrace();
+			}
+
 		}
-		
 		response.sendRedirect("/projetoweb/livros");
-		
-		
-		
 	}
 
 }
